audience components created
