+++ src/App.css
@@ -1,38 +1,130 @@
-.App {
+html  {
+  font-size: 20px;
+  box-sizing: border-box;
+  margin: 0 auto;
+
+  
+}
+
+*,
+*:before,
+*:after {
+  box-sizing: inherit;
+}
+
+header .img {
+
   text-align: center;
 }
+.nav{
 
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
+  
+  list-style:none;
+  margin:0;
+  padding:0;
+  text-align:center;
+}
+.nav li{
+  display:inline-block;
+}
+.nav a{
+  display:inline-block;
+  padding:10px;
 }
 
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
+
+.button {
+  border: solid 1px #000;
+  border-radius: 4px;
+  padding: 8px;
+  text-decoration: none;
+  line-height: 1;
+  
 }
 
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
+.solid {
+  background-color: #ff0000;
+  color: #FFF;
+}
+
+.hero {
+
+  display:flex;
+  
+}
+
+.call-to-action {
+
+  display:block;
+  flex-flow: row nowrap;
   align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
+  
+  background-size:contain; 
+  height: 394px;
+  width: 768px;
 }
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
+@media(max-width: 768px) {
+  html {
+      font-size: 16px;
+      box-sizing: border-box;
+      
   }
-  to {
-    transform: rotate(360deg);
+
+  *,
+  *:before,
+  *:after {
+      box-sizing: inherit;
   }
+
+  header .img {
+
+      text-align: center;
+  }
+  .nav{
+  
+      
+      list-style:none;
+      margin:0;
+      padding:0;
+      text-align:center;
+  }
+  .nav li{
+      display:inline-block;
+  }
+  .nav a{
+      display:inline-block;
+      padding:10px;
+  }
+
+
+  .button {
+      border: solid 1px #000;
+      border-radius: 4px;
+      padding: 8px;
+      text-decoration: none;
+      line-height: 1;
+      
+  }
+
+  .solid {
+      background-color: #ff0000;
+      color: #FFF;
+  }
+
+  .hero {
+
+      display:flex;
+      
+  }
+
+  .call-to-action {
+
+      display:block;
+      flex-flow: row nowrap;
+      align-items: center;
+      background-image: url("/dribble_concept.jpg");
+      background-size:contain; 
+      height: 394px;
+  }
+
 }
