+++ src/components/audience.jsx
@@ -1,0 +1,26 @@
+import React, { Component } from 'react';
+import { BrowserRouter as Router, Route } from 'react-router-dom';
+import AudienceJoinPres from './AudienceJoinPres'
+import "../App.css";
+
+class Audience extends Component {
+  
+  
+    render() {
+      
+      
+      return (
+      <>
+       <AudienceJoinPres/>
+      </>
+           
+                       
+      )
+      
+    }
+
+}
+
+
+export default Audience;
+
